Initial Commit (Working)
This tool generates a header file in C based on a picture which contains the ASCII glyphs of a font.
It's pretty self explanitory. Each pixel of each 8x16 glyph is encoded as a bit in the output array.
Data is written to a given output file. I've also included 2 fonts which I have created.